requirements.txt
asgiref==3.2.10
coverage==5.2.1
dj-database-url==0.5.0
Django==3.1
gunicorn==20.0.4
pytz==2020.1
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
django-rest-framework==0.1.0
pandas==1.4.1